app/requirements.txt
flask
flask_socketio
openai-whisper
librosa
#transformers 
faster_whisper
torch==2.3.0+cu121 
torchvision==0.18.0+cu121 
torchaudio==2.3.0 --index-url
